loadbalancer: Move experimental/benchmark to benchmark
