+++ pkg/loadbalancer/benchmark/testdata/endpointslice.yaml
